+++ src/components/RequestForm/RequestForm.tsx
@@ -1,8 +1,8 @@
 import { Moment } from "moment";
 import React, { useContext, useEffect, useState } from "react";
-import { Button, Col, Container, Form, Spinner } from "react-bootstrap";
+import { Button, Col, Container, Form, Spinner, Alert } from "react-bootstrap";
 import { Link, useHistory } from "react-router-dom";
-import { ApplicationTypes, Centers, IRequirementsRequestCRUD, OrgPriorities, RequirementsRequest, RequirementTypes } from "../../api/DomainObjects";
+import { ApplicationTypes, Centers, IRequirementsRequestCRUD, OrgPriorities, RequirementsRequest, RequirementTypes, IRequestValidation, RequestValidation } from "../../api/DomainObjects";
 import { IPerson, Person } from "../../api/UserApi";
 import { useScrollToTop } from "../../hooks/useScrollToTop";
 import { UserContext } from "../../providers/UserProvider";
@@ -10,6 +10,7 @@ import { CustomInputeDatePicker } from "../CustomInputDatePicker/CustomInputDate
 import { PeoplePicker } from "../PeoplePicker/PeoplePicker";
 import RequestSpinner from "../RequestSpinner/RequestSpinner";
 import './RequestForm.css';
+import moment from "moment";
 
 export interface IRequestFormProps {
     editRequestId?: number,
@@ -20,10 +21,12 @@ export interface IRequestFormProps {
 export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) => {
 
     const [request, setRequest] = useState<IRequirementsRequestCRUD>(new RequirementsRequest());
+    const [validation, setValidation] = useState<IRequestValidation | undefined>();
     const [showFundingField, setShowFundingField] = useState<boolean>(false);
     const [peoSameAsRequester, setPeoSameAsRequester] = useState<boolean>(false);
     const [saving, setSaving] = useState<boolean>(false);
     const [readOnly, setReadOnly] = useState<boolean>(false);
+    const [error, setError] = useState<string>("");
 
     const { user } = useContext(UserContext);
     const history = useHistory();
@@ -31,31 +34,42 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
     useScrollToTop();
 
     const getRequest = async () => {
-        if (props.editRequestId !== undefined && props.fetchRequestById) {
-            let newRequest = await props.fetchRequestById(props.editRequestId);
-            if (newRequest) {
-                setReadOnly(newRequest.isReadOnly());
-                setPeoSameAsRequester(newRequest.Requester.Id === newRequest.ApprovingPEO.Id);
-                setShowFundingField(newRequest.FundingOrgOrPEO !== "" && newRequest.FundingOrgOrPEO !== undefined && newRequest.FundingOrgOrPEO !== null);
-                setRequest(newRequest);
-            } else {
-                history.push("/Requests");
+        try {
+            if (props.editRequestId !== undefined && props.fetchRequestById) {
+                let newRequest = await props.fetchRequestById(props.editRequestId);
+                if (newRequest) {
+                    setReadOnly(newRequest.isReadOnly());
+                    setPeoSameAsRequester(newRequest.Requester.Id === newRequest.ApprovingPEO.Id);
+                    setShowFundingField(newRequest.FundingOrgOrPEO !== "" && newRequest.FundingOrgOrPEO !== undefined && newRequest.FundingOrgOrPEO !== null);
+                    setRequest(newRequest);
+                } else {
+                    history.push("/Requests");
+                }
             }
+        } catch (e) {
+            console.error("Error trying to fetch Request for Request Form");
+            console.error(e);
         }
     }
 
     // We need to update the state's request whenever the props.editRequest changes because the requests may not have loaded yet
     useEffect(() => {
         getRequest(); // eslint-disable-next-line
-    }, [props.editRequestId])
+    }, [props.editRequestId]);
 
     useEffect(() => {
         // only update the requester if this is a new request
         if (request.Id < 0) {
             updateRequest('Requester', user ? user : new Person({ Id: -1, Title: "Loading User", EMail: "" }));
-        }
-        // eslint-disable-next-line
-    }, [user])
+        } // eslint-disable-next-line
+    }, [user]);
+
+    useEffect(() => {
+        // Update validation whenever a field changes after a submission attempt
+        if (validation) {
+            setValidation(RequestValidation.getValidation(request, showFundingField));
+        } // eslint-disable-next-line
+    }, [request, showFundingField]);
 
     const updateRequest = (fieldUpdating: string, newValue: string | number | boolean | Moment | IPerson): void => {
         setRequest(new RequirementsRequest({ ...request, [fieldUpdating]: newValue }));
@@ -81,24 +95,53 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
     }
 
     const submitRequest = async (e: React.MouseEvent<HTMLButtonElement, MouseEvent> | React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        setSaving(true);
-        let req = request;
-        if (peoSameAsRequester) {
-            req.ApprovingPEO = req.Requester;
-            req.PEOOrgSymbol = req.RequesterOrgSymbol;
-            req.PEO_DSNPhone = req.RequesterDSNPhone;
-            req.PEO_CommPhone = req.RequesterCommPhone;
+        try {
+            e.preventDefault();
+            setSaving(true);
+            let req = request;
+            if (peoSameAsRequester) {
+                req.ApprovingPEO = req.Requester;
+                req.PEOOrgSymbol = req.RequesterOrgSymbol;
+                req.PEO_DSNPhone = req.RequesterDSNPhone;
+                req.PEO_CommPhone = req.RequesterCommPhone;
+            }
+            if (!showFundingField) {
+                req.FundingOrgOrPEO = "";
+            }
+            let requestValidation = RequestValidation.getValidation(req, showFundingField);
+            if (!requestValidation.IsErrored) {
+                await props.submitRequest(request);
+                history.push("/Requests");
+            } else {
+                setValidation(requestValidation);
+                setError("Please fix the errored fields!");
+            }
+        } catch (e) {
+            console.error("Error trying to submit request from request form");
+            console.error(e);
+        } finally {
+            setSaving(false);
         }
-        await props.submitRequest(request);
-        setSaving(false);
-        history.push("/Requests");
     }
 
     return (
         <Container className="pb-5 pt-3">
             <h1>{request.Id > -1 ? "Edit" : "New"} Request</h1>
             <Form className="request-form m-3" onSubmit={submitRequest}>
+                <Form.Row>
+                    <Col xl="3" lg="4" md="4" sm="4" xs="12">
+                        <CustomInputeDatePicker
+                            headerText="Requested Date:"
+                            readOnly={readOnly}
+                            date={request.RequestDate}
+                            maxDate={moment()}
+                            onChange={date => updateRequest('RequestDate', date)}
+                            isValid={validation && !validation.RequestDateError}
+                            isInvalid={validation && validation.RequestDateError !== ""}
+                            errorMessage={validation ? validation.RequestDateError : ""}
+                        />
+                    </Col>
+                </Form.Row>
                 <Form.Row>
                     <Col xl="6" lg="6" md="8" sm="12" xs="12">
                         <Form.Label>Requester:</Form.Label>
@@ -110,8 +153,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                                 updateRequest('Requester', persona ? new Person(persona) : new Person());
                             }}
                             readOnly={readOnly}
+                            isInvalid={validation && validation.RequesterError !== ""}
                             required
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.RequesterError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col xl="4" lg="4" md="4" sm="6" xs="12">
                         <Form.Label lg="4" sm="6">Requester Org Symbol:</Form.Label>
@@ -121,7 +168,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.RequesterOrgSymbol}
                             onChange={e => updateRequest('RequesterOrgSymbol', e.target.value)}
+                            isValid={validation && !validation.RequesterOrgSymbolError}
+                            isInvalid={validation && validation.RequesterOrgSymbolError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.RequesterOrgSymbolError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col xl="4" lg="4" md="6" sm="6" xs="12">
                         <Form.Label lg="4" sm="6">Requester DSN #:</Form.Label>
@@ -131,7 +183,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.RequesterDSNPhone}
                             onChange={e => updatePhoneField('RequesterDSNPhone', getNumbersOnly(e.target.value))}
+                            isValid={validation && !validation.RequesterDSNPhoneError}
+                            isInvalid={validation && validation.RequesterDSNPhoneError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.RequesterDSNPhoneError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col xl={{ span: 4, offset: 2 }} lg={{ span: 4, offset: 2 }} md="6" sm="6" xs="12">
                         <Form.Label lg="4" sm="6">Requester Comm #:</Form.Label>
@@ -141,7 +198,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.RequesterCommPhone}
                             onChange={e => updatePhoneField('RequesterCommPhone', getNumbersOnly(e.target.value))}
+                            isValid={validation && !validation.RequesterCommPhoneError}
+                            isInvalid={validation && validation.RequesterCommPhoneError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.RequesterCommPhoneError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col className="mt-4" xl="12" lg="12" md="12" sm="12" xs="12">
                         <Form.Check inline label="2 Ltr/PEO Same as Requester?" type="checkbox" id="peo-requester-checkbox"
@@ -163,7 +225,11 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             }}
                             readOnly={readOnly}
                             required
+                            isInvalid={validation && validation.ApprovingPEOError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.ApprovingPEOError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col xl="4" lg="4" md="4" sm="6" xs="12">
                         <Form.Label>2 Ltr/PEO Org Symbol:</Form.Label>
@@ -173,7 +239,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.PEOOrgSymbol}
                             onChange={e => updateRequest('PEOOrgSymbol', e.target.value)}
+                            isValid={validation && !validation.PEOOrgSymbolError}
+                            isInvalid={validation && validation.PEOOrgSymbolError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.PEOOrgSymbolError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col xl="4" lg="4" md="6" sm="6" xs="12">
                         <Form.Label>2 Ltr/PEO DSN #:</Form.Label>
@@ -183,7 +254,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.PEO_DSNPhone}
                             onChange={e => updatePhoneField('PEO_DSNPhone', getNumbersOnly(e.target.value))}
+                            isValid={validation && !validation.PEO_DSNPhoneError}
+                            isInvalid={validation && validation.PEO_DSNPhoneError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.PEO_DSNPhoneError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col xl={{ span: 4, offset: 2 }} lg={{ span: 4, offset: 2 }} md="6" sm="6" xs="12">
                         <Form.Label>2 Ltr/PEO Comm #:</Form.Label>
@@ -193,7 +269,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.PEO_CommPhone}
                             onChange={e => updatePhoneField('PEO_CommPhone', getNumbersOnly(e.target.value))}
+                            isValid={validation && !validation.PEO_CommPhoneError}
+                            isInvalid={validation && validation.PEO_CommPhoneError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.PEO_CommPhoneError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                 </Form.Row>}
                 <Form.Row>
@@ -205,7 +286,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.Title}
                             onChange={e => updateRequest('Title', e.target.value)}
+                            isValid={validation && !validation.TitleError}
+                            isInvalid={validation && validation.TitleError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.TitleError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col className="mt-4 mb-3" xl="12" lg="12" md="12" sm="12" xs="12">
                         <Form.Label className="mr-3 mb-0">Requirement Type:</Form.Label>
@@ -234,7 +320,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                                     readOnly={readOnly}
                                     value={request.FundingOrgOrPEO}
                                     onChange={e => updateRequest('FundingOrgOrPEO', e.target.value)}
+                                    isValid={validation && !validation.FundingOrgOrPEOError}
+                                    isInvalid={validation && validation.FundingOrgOrPEOError !== ""}
                                 />
+                                <Form.Control.Feedback type="invalid">
+                                    {validation ? validation.FundingOrgOrPEOError : ""}
+                                </Form.Control.Feedback>
                             </>
                         }
                     </Col>
@@ -261,7 +352,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                                     readOnly={readOnly}
                                     value={request.OtherApplicationNeeded}
                                     onChange={e => updateRequest('OtherApplicationNeeded', e.target.value)}
+                                    isValid={validation && !validation.OtherApplicationNeededError}
+                                    isInvalid={validation && validation.OtherApplicationNeededError !== ""}
                                 />
+                                <Form.Control.Feedback type="invalid">
+                                    {validation ? validation.OtherApplicationNeededError : ""}
+                                </Form.Control.Feedback>
                             </>
                         }
                     </Col>
@@ -296,7 +392,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.ProjectedOrgsImpactedOrg}
                             onChange={e => updateRequest('ProjectedOrgsImpactedOrg', e.target.value)}
+                            isValid={validation && !validation.ProjectedOrgsImpactedOrgError}
+                            isInvalid={validation && validation.ProjectedOrgsImpactedOrgError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.ProjectedOrgsImpactedOrgError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                 </Form.Row>
                 <Form.Row>
@@ -308,14 +409,23 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.ProjectedImpactedUsers ? request.ProjectedImpactedUsers : ''}
                             onChange={e => updateRequest("ProjectedImpactedUsers", parseInt(getNumbersOnly(e.target.value)))}
+                            isValid={validation && !validation.ProjectedImpactedUsersError}
+                            isInvalid={validation && validation.ProjectedImpactedUsersError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.ProjectedImpactedUsersError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                     <Col xl="3" lg="4" md="4" sm="4" xs="12">
                         <CustomInputeDatePicker
                             headerText="Operational Need Date:"
                             readOnly={readOnly}
                             date={request.OperationalNeedDate}
+                            minDate={moment()}
                             onChange={date => updateRequest('OperationalNeedDate', date)}
+                            isValid={validation && !validation.OperationalNeedDateError}
+                            isInvalid={validation && validation.OperationalNeedDateError !== ""}
+                            errorMessage={validation ? validation.OperationalNeedDateError : ""}
                         />
                     </Col>
                     <Col className="mt-4 mb-3" xl="12" lg="12" md="12" sm="12" xs="12">
@@ -356,7 +466,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.PriorityExplanation}
                             onChange={e => updateRequest('PriorityExplanation', e.target.value)}
+                            isValid={validation && !validation.PriorityExplanationError}
+                            isInvalid={validation && validation.PriorityExplanationError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.PriorityExplanationError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                 </Form.Row>
                 <Form.Row>
@@ -371,7 +486,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.BusinessObjective}
                             onChange={e => updateRequest('BusinessObjective', e.target.value)}
+                            isValid={validation && !validation.BusinessObjectiveError}
+                            isInvalid={validation && validation.BusinessObjectiveError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.BusinessObjectiveError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                 </Form.Row>
                 <Form.Row>
@@ -386,7 +506,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.FunctionalRequirements}
                             onChange={e => updateRequest('FunctionalRequirements', e.target.value)}
+                            isValid={validation && !validation.FunctionalRequirementsError}
+                            isInvalid={validation && validation.FunctionalRequirementsError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.FunctionalRequirementsError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                 </Form.Row>
                 <Form.Row>
@@ -401,7 +526,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.Benefits}
                             onChange={e => updateRequest('Benefits', e.target.value)}
+                            isValid={validation && !validation.BenefitsError}
+                            isInvalid={validation && validation.BenefitsError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.BenefitsError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                 </Form.Row>
                 <Form.Row>
@@ -416,7 +546,12 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                             readOnly={readOnly}
                             value={request.Risk}
                             onChange={e => updateRequest('Risk', e.target.value)}
+                            isValid={validation && !validation.RiskError}
+                            isInvalid={validation && validation.RiskError !== ""}
                         />
+                        <Form.Control.Feedback type="invalid">
+                            {validation ? validation.RiskError : ""}
+                        </Form.Control.Feedback>
                     </Col>
                 </Form.Row>
                 <Form.Row>
@@ -449,5 +584,19 @@ export const RequestForm: React.FunctionComponent<IRequestFormProps> = (props) =
                 </Link>
             </Form>
             <RequestSpinner show={saving} displayText="Saving Request..." />
+            {error &&
+                <Col className="mt-3 fixed-top"
+                    xl={{ span: 6, offset: 3 }}
+                    lg={{ span: 6, offset: 3 }}
+                    md={{ span: 8, offset: 2 }}
+                    sm={{ span: 10, offset: 1 }}
+                    xs={12}
+                >
+                    <Alert variant="danger" onClose={() => setError("")} dismissible>
+                        <Alert.Heading>Error Submitting!</Alert.Heading>
+                        <p>{error}</p>
+                    </Alert>
+                </Col>
+            }
         </Container>);
 }