+++ src/components/RequestRoutes/RequestRoutes.tsx
@@ -1,4 +1,5 @@
 import React, { useContext } from "react";
+import { Alert, Col } from "react-bootstrap";
 import { HashRouter, Redirect, Route, Switch } from "react-router-dom";
 import { useRequests } from "../../hooks/useRequests";
 import { UserContext } from "../../providers/UserProvider";
@@ -55,6 +56,20 @@ export const RequestRoutes: React.FunctionComponent<any> = (props) => {
             <RequestSpinner
                 show={requests.loading || loadingUser === undefined || loadingUser}
                 displayText={requests.loading ? "Loading Requests..." : "Loading User..."} />
+            {requests.error &&
+                <Col className="fixed-bottom"
+                    xl={{ span: 6, offset: 3 }}
+                    lg={{ span: 6, offset: 3 }}
+                    md={{ span: 8, offset: 2 }}
+                    sm={{ span: 10, offset: 1 }}
+                    xs={12}
+                >
+                    <Alert variant="danger" onClose={() => requests.clearError()} dismissible>
+                        <Alert.Heading>Error!</Alert.Heading>
+                        <p>{requests.error}</p>
+                    </Alert>
+                </Col>
+            }
         </>
     );
 }