+++ CHANGELOG.md
@@ -8,6 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 - (Keep your changes here until you have a release version)
 
+## [0.1.8] 2020-07-30
+### Added
+- Error handling on all external facing code
+- Error alerts for when errors do occur
+- Form validation on submit for the Request form
+- RequestDate field on form
+
 ## [0.1.7] 2020-07-28
 ### Added
 - Added Requester field that defaults to the current user.